+++ packages/botonic-react/src/webchat/hooks/use-webchat.ts
@@ -1,6 +1,7 @@
 import { Session } from '@botonic/core'
 import { useReducer, useRef } from 'react'
 
+import { Reply } from '../../components'
 import { ThemeProps, Webview } from '../../components/index-types'
 import { COLORS, WEBCHAT } from '../../constants'
 import { ClientInput, WebchatMessage } from '../../index-types'
@@ -68,11 +69,12 @@ export interface UseWebchat {
   updateLastRoutePath: (path: string) => void
   updateLatestInput: (input: ClientInput) => void
   updateMessage: (message: WebchatMessage) => void
-  updateReplies: (replies: any) => void
+  updateReplies: (replies: (typeof Reply)[]) => void
   updateSession: (session: Partial<Session>) => void
   updateTheme: (theme: ThemeProps, themeUpdates?: ThemeProps) => void
   updateTyping: (typing: boolean) => void
   updateWebview: (webview: Webview, params: Record<string, string>) => void
+  removeReplies: () => void
   removeWebview: () => void
   webchatState: WebchatState
   webchatRef: React.MutableRefObject<HTMLDivElement | null> // TODO: Change name, already exists WebchatRef for useImperativeHandle
@@ -108,8 +110,11 @@ export function useWebchat(): UseWebchat {
   const updateMessage = (message: WebchatMessage) =>
     webchatDispatch({ type: WebchatAction.UPDATE_MESSAGE, payload: message })
 
-  const updateReplies = replies =>
+  const updateReplies = (replies: (typeof Reply)[]) =>
     webchatDispatch({ type: WebchatAction.UPDATE_REPLIES, payload: replies })
+
+  const removeReplies = () =>
+    webchatDispatch({ type: WebchatAction.REMOVE_REPLIES, payload: [] })
 
   const updateLatestInput = (input: ClientInput) =>
     webchatDispatch({ type: WebchatAction.UPDATE_LATEST_INPUT, payload: input })
@@ -271,6 +276,7 @@ export function useWebchat(): UseWebchat {
     updateTheme,
     updateTyping,
     updateWebview,
+    removeReplies,
     removeWebview,
     webchatState,
     webchatRef,